fix: VoskBridge Recognizer 采样率转 Float（Kotlin 编译错误）
org.vosk.Recognizer 构造函数第二参为 float，SAMPLE_RATE 是 Int，
Kotlin 无隐式类型转换导致 compileReleaseKotlin 失败。

+++ android/app/src/main/kotlin/com/zqpd/wisemuse/VoskBridge.kt
@@ -93,7 +93,7 @@ class VoskBridge : FlutterPlugin, MethodChannel.MethodCallHandler {
         val bufferSize = AudioRecord.getMinBufferSize(
             SAMPLE_RATE, AudioFormat.CHANNEL_IN_MONO, AudioFormat.ENCODING_PCM_16BIT
         )
-        recognizer = Recognizer(model, SAMPLE_RATE)
+        recognizer = Recognizer(model, SAMPLE_RATE.toFloat())
         audioRecord = AudioRecord(
             MediaRecorder.AudioSource.MIC,
             SAMPLE_RATE,
